refactor(core): Move frontend settings to @n8n/api-types (no-changelog)
