+++ cypress/package.json
@@ -15,6 +15,7 @@
     "start": "cd ..; pnpm start"
   },
   "devDependencies": {
+    "@n8n/api-types": "workspace:*",
     "@types/lodash": "catalog:",
     "eslint-plugin-cypress": "^3.3.0",
     "n8n-workflow": "workspace:*"
