+++ packages/@n8n/api-types/src/index.ts
@@ -1,6 +1,7 @@
+export type * from './datetime';
 export type * from './push';
 export type * from './scaling';
-export type * from './datetime';
+export type * from './frontend-settings';
 export type * from './user';
 
 export type { Collaborator } from './push/collaboration';
