+++ packages/editor-ui/src/__tests__/defaults.ts
@@ -1,6 +1,6 @@
-import type { IN8nUISettings } from 'n8n-workflow';
+import type { FrontendSettings } from '@n8n/api-types';
 
-export const defaultSettings: IN8nUISettings = {
+export const defaultSettings: FrontendSettings = {
 	databaseType: 'sqlite',
 	isDocker: false,
 	pruning: {
