+++ packages/editor-ui/src/components/Telemetry.vue
@@ -1,8 +1,8 @@
 <script lang="ts" setup>
+import type { ITelemetrySettings } from '@n8n/api-types';
 import { useRootStore } from '@/stores/root.store';
 import { useSettingsStore } from '@/stores/settings.store';
 import { useUsersStore } from '@/stores/users.store';
-import type { ITelemetrySettings } from 'n8n-workflow';
 import { useProjectsStore } from '@/stores/projects.store';
 import { computed, onMounted, watch, ref } from 'vue';
 import { useTelemetry } from '@/composables/useTelemetry';
