+++ packages/editor-ui/src/stores/settings.store.ts
@@ -1,5 +1,6 @@
 import { computed, ref } from 'vue';
 import Bowser from 'bowser';
+import type { IUserManagementSettings, FrontendSettings } from '@n8n/api-types';
 
 import * as publicApiApi from '@/api/api-keys';
 import * as ldapApi from '@/api/ldap';
@@ -8,12 +9,7 @@ import { testHealthEndpoint } from '@/api/templates';
 import type { ILdapConfig } from '@/Interface';
 import { STORES, INSECURE_CONNECTION_WARNING } from '@/constants';
 import { UserManagementAuthenticationMethod } from '@/Interface';
-import type {
-	IDataObject,
-	IN8nUISettings,
-	WorkflowSettings,
-	IUserManagementSettings,
-} from 'n8n-workflow';
+import type { IDataObject, WorkflowSettings } from 'n8n-workflow';
 import { ExpressionEvaluatorProxy } from 'n8n-workflow';
 import { defineStore } from 'pinia';
 import { useRootStore } from './root.store';
@@ -27,7 +23,7 @@ import { i18n } from '@/plugins/i18n';
 
 export const useSettingsStore = defineStore(STORES.SETTINGS, () => {
 	const initialized = ref(false);
-	const settings = ref<IN8nUISettings>({} as IN8nUISettings);
+	const settings = ref<FrontendSettings>({} as FrontendSettings);
 	const userManagement = ref<IUserManagementSettings>({
 		quota: -1,
 		showSetupOnFirstLoad: false,
@@ -164,7 +160,7 @@ export const useSettingsStore = defineStore(STORES.SETTINGS, () => {
 
 	const isDevRelease = computed(() => settings.value.releaseChannel === 'dev');
 
-	const setSettings = (newSettings: IN8nUISettings) => {
+	const setSettings = (newSettings: FrontendSettings) => {
 		settings.value = newSettings;
 		userManagement.value = newSettings.userManagement;
 		if (userManagement.value) {
@@ -208,7 +204,7 @@ export const useSettingsStore = defineStore(STORES.SETTINGS, () => {
 		}
 	};
 
-	const setAllowedModules = (allowedModules: IN8nUISettings['allowedModules']) => {
+	const setAllowedModules = (allowedModules: FrontendSettings['allowedModules']) => {
 		settings.value.allowedModules = allowedModules;
 	};
 
@@ -367,7 +363,7 @@ export const useSettingsStore = defineStore(STORES.SETTINGS, () => {
 	};
 
 	const reset = () => {
-		settings.value = {} as IN8nUISettings;
+		settings.value = {} as FrontendSettings;
 	};
 
 	return {
