--- packages/workflow/src/Authentication.ts
@@ -1,1 +1,0 @@
-export type AuthenticationMethod = 'email' | 'ldap' | 'saml';
