+++ packages/workflow/src/index.ts
@@ -6,7 +6,6 @@ import * as ObservableObject from './ObservableObject';
 import * as TelemetryHelpers from './TelemetryHelpers';
 
 export * from './errors';
-export * from './Authentication';
 export * from './Constants';
 export * from './Cron';
 export * from './DeferredPromise';
